+++ CHANGES.md
@@ -26,6 +26,7 @@ Apollo 2.0.0
 * [Catch LinkageError for ClassLoaderUtil.isClassPresent in case class is present but is failed to load](https://github.com/apolloconfig/apollo/pull/4097)
 * [Split helm chart into another repo](https://github.com/apolloconfig/apollo/pull/4125)
 * [fix gray publish refresh item status](https://github.com/apolloconfig/apollo/pull/4128)
+* [Support only show difference keys when compare namespace](https://github.com/apolloconfig/apollo/pull/4165)
 
 ------------------
 All issues and pull requests are [here](https://github.com/ctripcorp/apollo/milestone/8?closed=1)